Updated web services

+++ eg-backend/Ethics-Learning/src/main/java/com/mahindra/eng/infrastructure/entity/EthicsCounselor.java
@@ -20,9 +20,9 @@ public class EthicsCounselor implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int locationId;
+    private Long locationId;
     private int employeeId;
-    private int companyId;
+    private Long companyId;
     private String type;
     private String createdBy;
     @CreatedDate
